requirements.txt
requests>=2.32.3
supabase>=2.11.0
python-dotenv 